requirements.txt
aiofiles==25.1.0
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
cachetools==6.2.1
cassandra-driver==3.29.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
cqlsh==6.2.1
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
emoji==2.15.0
executing==2.2.1
filetype==1.2.0
frozenlist==1.8.0
geomet==0.2.1.post1
google-ai-generativelanguage==0.6.6
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hiredis==3.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-openai==0.2.0
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.2
Markdown==3.9
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==1.26.4
olefile==0.47
openai==1.109.1
orjson==3.11.3
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.1
ptyprocess==0.7.0
pure-sasl==0.6.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
redis==5.2.1
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.15
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==2.0.0
yarl==1.22.0
